% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.r
\name{CoCoNet}
\alias{CoCoNet}
\title{The main function that leverages gene co-expression pattern to infer trait-relevant tissues in genome-wide association studies}
\usage{
CoCoNet(gene_effect_size, max_path = 1, adjacency_mat)
}
\arguments{
\item{gene_effect_size:}{A length m vector of gene level effect sizes for m genes}

\item{max_path:}{maximum number of paths linking between two genes considered in the model. The default value is 1.}

\item{adjacency_mat:}{A m by m adjacency matrix}
}
\value{
A list of estimated parameters
\item{loglikelihood}{A value of log likelihood of selected disease-tissue pair}
\item{sigma0}{The estimated sigma0 square in our model, where sigma0 square multiply adjacency_mat measures gene-level effect size correlation due to direct connections among genes}
\item{sigma1}{The estimated sigma1 square in our model, where sigma1 square multiply adjacency_mat measures gene-level effect size correlation due to indirect connections among genes through two-paths}
\item{sigma2}{If A2 model is selected, this is the estimated sigma2 square in our model, where sigma2 square multiply adjacency_mat measures gene-level effect size correlation due to indirect connections among genes through two-paths}
\item{sigma3}{If A3 model is selected, this is the estimated sigma3 square in our model, where sigma3 square multiply adjacency_mat measures gene-level effect size correlation due to indirect connections among genes through three-paths}
\item{sigma4}{If A4 model is selected, this is the estimated sigma4 square in our model, where sigma4 square multiply adjacency_mat measures gene-level effect size correlation due to indirect connections among genes through four-paths}
\item{rho}{A value measuring relative signal strength of one-path gene co-expression pattern on gene-level effect sizes}
}
\description{
CoCoNet applies a composite likelihood-based covariance regression network model to identify trait-relevant tissues or cell types
}
